fix: address code review findings on lint suppression removal
- formatPrompt: look up the provider params block by the `type`
  discriminant instead of taking the first non-`type` key
- dedupe parseExperimentDatum into ExperimentMetricsTooltipHeader
- LDAPLoginForm: await navigate to satisfy no-floating-promises
